3-06

+++ src/components/order/cart-items.tsx
@@ -201,23 +201,25 @@ export function CartItems({
                                     })}
                                     {
                                         place.charges.map((charge: any) => {
-                                            return (
-                                                <div className="flex justify-between" key={charge.id}>
-                                                    <div className="overflow-hidden grow-1 my-col-5">
-                                                        <div className="truncate ">
-                                                            {charge.name}
+                                            if(charge.amount > 0){
+                                                return (
+                                                    <div className="flex justify-between" key={charge.id}>
+                                                        <div className="overflow-hidden grow-1 my-col-5">
+                                                            <div className="truncate ">
+                                                                {charge.name}
+                                                            </div>
+                                                        </div>
+                                                        <div className="flex flex-col grow-0 items-end gap-y-1.25">
+                                                            <div>
+                                                                {formatPrice(
+                                                                    charge.amount
+                                                                )}{" "}
+                                                                ₮
+                                                            </div>
                                                         </div>
                                                     </div>
-                                                    <div className="flex flex-col grow-0 items-end gap-y-1.25">
-                                                        <div>
-                                                            {formatPrice(
-                                                                charge.amount
-                                                            )}{" "}
-                                                            ₮
-                                                        </div>
-                                                    </div>
-                                                </div>
-                                            )
+                                                )
+                                            }
                                         })
                                     }
                                 </div>
@@ -230,9 +232,9 @@ export function CartItems({
                             <div>Захиалгын дүн:</div>
                             {deliveryType == "Delivery" ? (
                                 <>
-                                    <div>Хүргэлтийн төлбөр:</div>
+                                    <div>Хүргэлтийн төлбөр: </div>
                                     {
-                                        (packageAmount && packageAmount) !== 0 && (
+                                        packageAmount !== 0 && (
                                             <div>Савны мөнгө:</div>
                                         )
                                     }
@@ -248,7 +250,7 @@ export function CartItems({
                                 <>
                                     <div>{formatPrice(taxAmount)} ₮</div>
                                     {
-                                        (packageAmount && packageAmount) !== 0 && (
+                                        packageAmount !== 0 && (
                                             <div>{formatPrice(packageAmount)} ₮</div>
                                         )
                                     }
